move helper function and fix bug
